1ER ENTREGA ORD 24032026

+++ src/modules/datestord/datestord.entity.ts
@@ -2,7 +2,7 @@ import { Column, Entity, PrimaryColumn } from 'typeorm';
 
 @Entity({ name: 'DAT_EST_ORD', schema: 'dbo' })
 export class DatEstOrdEntity {
-  @PrimaryColumn({ name: 'ESTA', type: 'int' })
+  @PrimaryColumn({ name: 'ESTA', type: 'float' })
   ESTA: number;
 
   @Column({ name: 'TIPO', type: 'nvarchar', length: 255, nullable: true })
